Server/BackFlask/requirements.txt
flask >= 3.0.0
psycopg2-binary
python-dotenv >= 1.0.0
flask-jwt-extended >= 4.6.0